% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df_format.R
\name{check_df_format}
\alias{check_df_format}
\title{Checks the format of a data frame}
\usage{
check_df_format(df, format = NULL, dev_mode = FALSE)
}
\arguments{
\item{df}{query data frame}

\item{format}{format to test - see \code{\link{orgsurveyr-data-formats}}}

\item{dev_mode}{stop if test is false.  Default is FALSE}
}
\value{
string of data frame type
}
\description{
Helper function to determine whether a supplied data frame is a specific \code{\link{orgsurveyr-data-formats}}
defined in the package.  Either returns a boolean or throws a stop.
}
\examples{
data(tg_org_indiv_df)
check_df_format(tg_org_indiv_df, 'indiv_df')

data(tg_org_indiv_tall_df)
check_df_format(tg_org_indiv_tall_df, 'indiv_tall_df')

data(tg_org_summarised_df)
check_df_format(tg_org_summarised_df, 'org_tall_df')

data(mtcars)
check_df_format(mtcars, 'indiv_df')
}
