% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regular_org.R
\name{create_regular_org}
\alias{create_regular_org}
\title{Create a symmetrical organisation}
\usage{
create_regular_org(n_children = 4, max_depth = 3)
}
\arguments{
\item{n_children}{integer defining the number of children each node should have}

\item{max_depth}{the maximum depth of the organisation}
}
\value{
tidygraph object
}
\description{
Create an organisation where each unit has a specified number of child units up until
a maximum depth.
}
\examples{
set.seed(1234)
tg1a <- create_regular_org(n_children=4, max_depth=3)
tg1a

\dontrun{
plot_org(tg1a, fill_var=NULL)
}
}
