% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_format.R
\name{get_df_format}
\alias{get_df_format}
\title{Determine format of data frame}
\usage{
get_df_format(df)
}
\arguments{
\item{df}{query data frame}
}
\value{
string of data frame type
}
\description{
Helper function to determine whether a supplied data frame is of any of the
\code{\link{orgsurveyr-data-formats}} defined in the package.
}
\examples{
data(tg_org_indiv_df)
get_df_format(tg_org_indiv_df)

data(tg_org_indiv_minimal_df)
get_df_format(tg_org_indiv_minimal_df)

data(tg_org_indiv_tall_df)
get_df_format(tg_org_indiv_tall_df)

data(tg_org_summarised_df)
get_df_format(tg_org_summarised_df)

data(mtcars)
get_df_format(mtcars)
}
