% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shiny_app.R
\name{make_shiny_app}
\alias{make_shiny_app}
\title{Create an example shiny app}
\usage{
make_shiny_app(path = "~")
}
\arguments{
\item{path}{Location where the file should be created. Defaults to the user's home directory}
}
\value{
creates an example shiny app file
}
\description{
Creates an example shiny application file which can be used as a template to create
shiny apps with different configurations for deployment on a shiny server.
}
\examples{
\dontrun{
# create the example shiny app
make_shiny_app()

# then edit and copy to the correct location
mv ~/example_shiny_app.R ~/ShinyApps/my_app/app.R
}
}
