% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgsurveyr-package.r
\docType{package}
\name{orgsurveyr-package}
\alias{orgsurveyr-package}
\alias{orgsurveyr}
\title{Analyise and visualise organisation data}
\description{
Lightweight wrapper for \pkg{ggraph} and \pkg{tidygraph} providing vignettes
and convenience functions for the analysis of organisational data.
}
\details{
To look at the package vignettes type \code{ browseVignettes('orgsurveyr')}
}
\author{
Phil Chapman <phil.chapman@dwp.gsi.gov.uk>
}
