% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_orgviz.R
\name{orgviz}
\alias{orgviz}
\title{Orgviz shiny app}
\usage{
orgviz(tg = NULL, df = NULL)
}
\arguments{
\item{tg}{user supplied tbl_graph}

\item{df}{user supplied a data frame in org_tall_df format
ie summarised metrics for the units in the network as generated by calc_summary_df.}
}
\value{
runs a shiny app
}
\description{
Orgviz shiny app
}
\examples{
NULL
}
