% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_org.R
\name{plot_org}
\alias{plot_org}
\title{Simple organisation plot}
\usage{
plot_org(x, fill_var = "depth", is_circular = FALSE)
}
\arguments{
\item{x}{organisation as a tidygraph object (tidygraph)}

\item{fill_var}{the variable to use as a fill colour for units (column name)}

\item{is_circular}{whether to make the dendrogram circular which is useful for larger organisations (logical)}
}
\value{
ggraph object
}
\description{
Bare bones function to plot an organisation style dendrogram from a tidygraph object.
}
\details{
Note that the recommended way to plot organisations is using the ggraph package.  Some examples of how to do this
are included in the vignettes.  This simple function is provided as a convenience for new users and to use in examples.

DEV NOTE: in ggplot2 version 3.0.0 tidyeval is available so should be able to use this rather than aes_string
}
\examples{
\dontrun{
set.seed(1234)
tg <- create_realistic_org()
plot_org(tg)
plot_org(tg, is_circular=TRUE)
}
}
