% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_individuals_df.R
\name{simulate_individuals_df}
\alias{simulate_individuals_df}
\title{Simulate individuals data frame}
\usage{
simulate_individuals_df(x)
}
\arguments{
\item{x}{tbl_graph organisation with unit_size defined}
}
\value{
tibble
}
\description{
Given an organisation tbl_graph object with the unit_size column defined, a tibble will be generated
with one row per individual in the organisation.  For test purposes a dummy variable is also generated.
}
\examples{
set.seed(1234)
tg_ex1 <- create_realistic_org(n_children = 4, max_depth = 3, prob=0.3)
tg_ex1 <- simulate_unit_size(tg_ex1)
df <- simulate_individuals_df(tg_ex1)
df
}
