% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_test.R
\name{test_viz}
\alias{test_viz}
\title{Testviz shiny app}
\usage{
test_viz()
}
\value{
runs a shiny app
}
\description{
Convenience function to run the test_viz shiny app within an R session.
}
\details{
The test_viz shiny app allows interactive simulation and visualisation of organisations of different shapes and
sizes.
}
\examples{
\dontrun{

test_viz()

}
}
