% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_test.R
\name{test_viz_server}
\alias{test_viz_server}
\title{Testviz shiny server}
\usage{
test_viz_server(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\value{
shiny server function
}
\description{
Creates a shiny server object for the test_viz shiny app
}
\examples{
\dontrun{
library(shiny)

shinyApp(ui = test_viz_ui(),
        server = function(input, output) {
          test_viz_server(input, output)
        }
)
}
}
