% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_test.R
\name{test_viz_ui}
\alias{test_viz_ui}
\title{Testviz shiny UI}
\usage{
test_viz_ui()
}
\value{
shiny ui object
}
\description{
Creates a shiny UI object for the test_viz shiny app
}
\examples{
\dontrun{
library(shiny)

shinyApp(ui = test_viz_ui(),
        server = function(input, output) {
          test_viz_server(input, output)
        }
)
}
}
