% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgsurveyr-data-examples.R
\docType{data}
\name{tg_org}
\alias{tg_org}
\title{Example organisation tbl_graph object}
\format{An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.}
\usage{
tg_org
}
\description{
TO ADD
}
\keyword{datasets}
