% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgsurveyr-data-examples.R
\docType{data}
\name{tg_org_summarised_df}
\alias{tg_org_summarised_df}
\title{Example org_tall_df data frame}
\format{A data frame with 76 rows and 3 variables:
\describe{
 \item{unit_id}{The organisational unit}
 \item{metric_id}{The name of the variable}
 \item{value}{The value of the variable}
 }}
\usage{
tg_org_summarised_df
}
\description{
An example of a data frame in the org_tall_df data frame format where
an organisational unit is represented by multiple rows, and each row represents a
different summarised variable.
}
\keyword{datasets}
