% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_df.R
\name{calc_summary_df}
\alias{calc_summary_df}
\title{Calculate summary metrics on an organisation}
\usage{
calc_summary_df(tg, df, tall_df = NULL, selected_vars,
  is_cumulative = FALSE)
}
\arguments{
\item{tg}{tbl_graph that passes a check with check_tbl_graph_is_org}

\item{df}{a data frame with columns named unit_id and invididual_id and
(and optionally individual level variables) with one row per individual}

\item{tall_df}{a data frame with columns named individual id, metric_id and value with multiple rows per individual}

\item{selected_vars}{names of variables in df or tall_df (character vector)}

\item{is_cumulative}{whether to calculate cumulative or orthodox aggregations (logical)}
}
\value{
A tall data frame with three columns: unit_id, metric_id and value.
}
\description{
Given individual level variables and an organisation structure, this function calculates aggregated
metrics using either the cumulative approach (all individuals in that unit or its descendents) or the orthodox
approach (individuals immediately associated with that unit only).
}
\examples{
NULL
}
