% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_graph_is_org.R
\name{check_tbl_graph_is_org}
\alias{check_tbl_graph_is_org}
\title{Check tbl_graph is an organisation}
\usage{
check_tbl_graph_is_org(x)
}
\arguments{
\item{x}{}
}
\value{
logical
}
\description{
Verify that a tbl_graph has the right features to be an organisation
}
\examples{

library(tidygraph)

set.seed(1234)
tg1 <- create_realistic_org()
check_tbl_graph_is_org(tg1)
# returns TRUE

tg2 <- tidygraph::create_star(30)
check_tbl_graph_is_org(tg2)
# returns FALSE

}
