% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_realistic_org.R
\name{create_realistic_org}
\alias{create_realistic_org}
\title{Create a realistic organisation}
\usage{
create_realistic_org(n_children = 4, max_depth = 3, prob = 0.3,
  .f = NULL, keep_creation_vars = FALSE)
}
\arguments{
\item{n_children}{integer}

\item{max_depth}{integer}

\item{prob}{double or vector of doubles}

\item{.f}{function of x}

\item{keep_creation_vars}{logical}
}
\value{
tidygraph object
}
\description{
Create an organisation where each unit starts with a specified number of child units up until
a maximum depth, but where units are then randomly removed according to probabilities set in either
a single value for all units, a value per unit depth or a user defined function of depth.
}
\examples{
tg1 <- create_realistic_org(4,3, prob=0.3)
tg1

tg2 <- create_realistic_org(4,3, prob=c(0.2, 0.4, 0.6))
tg2
}
