% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_realistic_org.R
\name{create_realistic_org}
\alias{create_realistic_org}
\title{Create a realistic organisation}
\usage{
create_realistic_org(n_children = 4, max_depth = 3, prob = 0.3,
  .f = NULL, delete_units = TRUE)
}
\arguments{
\item{n_children}{number of child units per parent unit (integer)}

\item{max_depth}{integer maximum depth of organisation (integer)}

\item{prob}{probability that a unit is deleted to create an irregular organisation.
If a vector is provided that is the same length as max_depth then a different probability can be assigned to different
levels of the organisation (double or vector of doubles)}

\item{.f}{function of depth that generates probability that a unit is removed (NOT IMPLEMENTED) (function of x)}

\item{delete_units}{retains intermediate variables and doesn't delete units - useful to understand how the function
is working (logical)}
}
\value{
tidygraph object
}
\description{
Create an organisation where each unit starts with a specified number of child units up until
a maximum depth, but where units are then randomly removed according to probabilities set in either
a single value for all units, a value per unit depth or a user defined function of depth.
}
\examples{
set.seed(1234)
tg1 <- create_realistic_org(4,3, prob=0.3)
tg1

set.seed(1234)
tg2 <- create_realistic_org(4,3, prob=c(0.2, 0.4, 0.6))
tg2
}
