% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cumulative_mapping_df.R
\name{generate_cumulative_mapping}
\alias{generate_cumulative_mapping}
\title{Generate Cumulative Mapping}
\usage{
generate_cumulative_mapping(tg, df)
}
\arguments{
\item{tg}{tbl_graph that passes a check with check_tbl_graph_is_org}

\item{df}{a data frame with columns named unit_id and invididual_id}
}
\value{
tbl_df
}
\description{
Takes an organisation as a tbl_graph object and a data frame that maps individuals to units
and returns a data frame that maps an organisational unit to all individuals within that unit.
}
\details{
For example, if an organisation is split into four departments, and each department has four teams,
an employee is affiliated not just to their immediate team, but also their department and the organisation
as a whole.

The output dataframe serves as a basis for calculating unit level summaries that include all individuals
affiliated to that unit.
}
\examples{
NULL
}
