% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_orgviz.R
\name{orgviz}
\alias{orgviz}
\title{Orgviz shiny app}
\usage{
orgviz(tg = NULL, df = NULL)
}
\arguments{
\item{tg}{user supplied tbl_graph}

\item{df}{user supplied a data frame in org_tall_df format
ie summarised metrics for the units in the network as generated by calc_summary_df.}
}
\value{
runs a shiny app
}
\description{
Convenience function to run the orgviz shiny app within an R session.
}
\details{
The orgviz shiny app permits interactive exploration of an organisational network, and the mapping
of numeric data onto the organisation.  By default a dataset is simulated, but a tidygraph object and
a data frame in the org_tall_df format call be supplied to visualise custom data.
}
\examples{
\dontrun{

# run with embedded data
orgviz()

# run with example data
orgviz(tg=orgsurveyr::tg_org, df=orgsurveyr::tg_org_summarised_df)

}
}
