% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_orgviz.R
\name{orgviz_server}
\alias{orgviz_server}
\title{Orgviz shiny server}
\usage{
orgviz_server(input, output, tg = NULL, df = NULL)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{tg}{user supplied tbl_graph}

\item{df}{user supplied a data frame in org_tall_df format
ie summarised metrics for the units in the network as generated by calc_summary_df.}
}
\value{
shiny server function
}
\description{
Orgviz shiny server
}
\examples{
NULL
}
