% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_orgviz.R
\name{orgviz_server}
\alias{orgviz_server}
\title{Orgviz shiny server}
\usage{
orgviz_server(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\value{
shiny server function
}
\description{
Orgviz shiny server
}
\examples{
NULL
}
