% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_unit_size.R
\name{simulate_unit_size}
\alias{simulate_unit_size}
\title{Simulate unit sizes within an organisation}
\usage{
simulate_unit_size(x)
}
\arguments{
\item{x}{a tbl_graph object representing an organisation}
}
\description{
Function which takes a tbl_graph object representing an organisation and generates a number representing
the number of individuals immediately associated each unit.
}
\examples{
NULL
}
