% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{test_viz_server}
\alias{test_viz_server}
\title{Test visualisation server}
\usage{
test_viz_server(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\value{
shiny server function
}
\description{
Test visualisation server
}
\examples{
NULL
}
