% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgsurveyr-data-examples.R
\docType{data}
\name{tg_org_summarised_df}
\alias{tg_org_summarised_df}
\title{Example org_tall_df data frame}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 76 rows and 3 columns.}
\usage{
tg_org_summarised_df
}
\description{
TO ADD
}
\keyword{datasets}
